--- .eslintrc
@@ -1,20 +1,0 @@
-{
-    "extends": ["prettier-standard"],
-    "plugins": ["mocha", "chai-friendly", "prettier"],
-    "rules": {
-        "prettier/prettier": ["error", { "tabWidth": 4 }],
-        "mocha/no-exclusive-tests": "error",
-        "chai-friendly/no-unused-expressions": 2
-    },
-    "env": {
-        "es6": true,
-        "mocha": true,
-        "node": true
-    },
-    "globals": {
-        "artifacts": true,
-        "contract": true,
-        "assert": true,
-        "web3": true
-    }
-}
