+++ test/sol6/fuzzerFiles/randomNumberGenerator.js
