+++ test/sol6/fuzzerFiles/trades/multiple_trade_fuzz_test.sh
