+++ test/sol6/fuzzerFiles/trades/tradeParamsGenerator.js
@@ -1,6 +1,6 @@
 const BN = web3.utils.BN;
-const Helper = require("../../test/helper.js");
-const nwHelper = require("../../test/sol6/networkHelper.js");
+const Helper = require("../../../helper.js");
+const nwHelper = require("../../networkHelper.js");
 
 require("chai")
     .use(require("chai-as-promised"))
@@ -8,8 +8,8 @@ require("chai")
     .should();
 
 
-const {BPS, ethDecimals, ethAddress, emptyHint, zeroBN, zeroAddress }  = require("../../test/helper.js")
-const { MASK_IN_HINTTYPE, MASK_OUT_HINTTYPE, SPLIT_HINTTYPE, BEST_OF_ALL_HINTTYPE }  = require('../../test/sol6/networkHelper.js');
+const {BPS, ethDecimals, ethAddress, emptyHint, zeroBN, zeroAddress }  = require("../../../helper.js")
+const { MASK_IN_HINTTYPE, MASK_OUT_HINTTYPE, SPLIT_HINTTYPE, BEST_OF_ALL_HINTTYPE }  = require('../../networkHelper.js');
 
 // Operations
 const TRADE = 'trade';
